refactor(ds): update Icon component - remove redundant code
